Finished rewriting the bake logic without the prototype.
The rewrite turned out to be a major pain in the ass and clusterfuck
at the same time. I did not expect to run into *so* many corner cases
when rewriting this logic. Additionally some of my previous assumptions
turned out to be a little wrong...it was a pain to make it short.
I think I spend over 6 or 8 hours now only rewriting this part...